init refactor
